pyproject.toml
[tool.poetry]
name = "netmanageit-connector"
version = "0.1.0"
description = ""
authors = ["ix-bahy <m.elbahy@buguard.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pycti = "6.2.18"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
